+++ CHANGELOG.md
@@ -32,6 +32,9 @@
 * Support user state. ([#525](https://github.com/realm/realm-dart/pull/525))
 * Support getting user id and identities. ([#525](https://github.com/realm/realm-dart/pull/525))
 * Support user logout. ([#525](https://github.com/realm/realm-dart/pull/525))
+* Support user deviceId. ([#570](https://github.com/realm/realm-dart/pull/570))
+* Support user authentication provider type. ([#570](https://github.com/realm/realm-dart/pull/570))
+* Support user profile data. ([#570](https://github.com/realm/realm-dart/pull/570))
 
 ### Fixed
 * Fixed an issue that would result in the wrong transaction being rolled back if you start a write transaction inside a write transaction. ([#442](https://github.com/realm/realm-dart/issues/442))